[class] feat : Phase 3+4 — 클래스 라우트 vibe-intro → [slug] 동적 세그먼트 전환
(class-lesson)/class/vibe-intro/lesson/[id] → /class/[slug]/lesson/[id]
(landing)/class/vibe-intro/** → /class/[slug]/**
[id]/page.tsx를 [slug]/page.tsx로 통합 (동일 레벨 동적 세그먼트 충돌 해결)
(service)/payment: courseSlug 기반 폴백 처리
