+++ src/app/(class-lesson)/class/[slug]/lesson/[id]/_components/lesson-tabs.tsx
