+++ src/app/(landing)/class/[slug]/(learning)/feed/_components/paywall-section.tsx
