+++ src/app/(landing)/class/[slug]/(learning)/feed/page.tsx
@@ -1,0 +1,10 @@
+import { redirect } from 'next/navigation';
+
+export default async function FeedIndexPage({
+  params,
+}: {
+  params: Promise<{ slug: string }>;
+}) {
+  const { slug } = await params;
+  redirect(`/class/${slug}/home?tab=feed`);
+}
