+++ src/app/(landing)/class/[slug]/(learning)/qa/page.tsx
@@ -1,0 +1,10 @@
+import { redirect } from 'next/navigation';
+
+export default async function QnaIndexPage({
+  params,
+}: {
+  params: Promise<{ slug: string }>;
+}) {
+  const { slug } = await params;
+  redirect(`/class/${slug}/home?tab=qna`);
+}
