--- src/app/(landing)/class/vibe-intro/(learning)/feed/[id]/edit/page.tsx
@@ -1,19 +1,0 @@
-'use client';
-
-import { useRouter } from 'next/navigation';
-import { use, useEffect } from 'react';
-
-export default function FeedEditPage({
-  params,
-}: {
-  params: Promise<{ id: string }>;
-}): null {
-  const { id } = use(params);
-  const router = useRouter();
-
-  useEffect(() => {
-    router.replace(`/class/vibe-intro/feed/${id}`);
-  }, [id, router]);
-
-  return null;
-}
