--- src/app/(landing)/class/vibe-intro/(learning)/feed/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default function FeedIndexPage() {
-  redirect('/class/vibe-intro/home?tab=feed');
-}
