--- src/app/(landing)/class/vibe-intro/(learning)/qa/page.tsx
@@ -1,5 +1,0 @@
-import { redirect } from 'next/navigation';
-
-export default function QnaIndexPage() {
-  redirect('/class/vibe-intro/home?tab=qna');
-}
